Move Dummy string repr to separate utils file
