+++ torch_ttnn/passes/add_data_move_pass.py
@@ -1,4 +1,10 @@
 import torch
+from ..utils import (
+    DummyTtnnUint32,
+    DummyTtnnRowMajorLayout,
+    DummyTtnnTileLayout,
+    DummyDevice,
+)
 
 try:
     import ttnn
@@ -148,26 +154,6 @@ def try_add_data_move_out_for_split(src_node, dst_idx, dst_node) -> torch.fx.nod
     else:
         return None
 
-# See https://docs.google.com/document/d/1r2D4AagoeTRjEmXFnWzzafaWQkf-8hlIbX2ze-JAUFo/edit#heading=h.zad9rwqjv6cr
-class DummyDevice:
-    def __repr__(self):
-        return f"ttnn_Specified_Device"
-
-class DummyTtnnRowMajorLayout:
-    def __repr__(self):
-        return f"ttnn_ROW_MAJOR_LAYOUT"
-
-class DummyTtnnTileLayout:
-    def __repr__(self):
-        return f"ttnn_TILE_LAYOUT"
-
-class DummyTtnnUint32:
-    def __repr__(self):
-        return f"ttnn_uint32"
-
-class DummyTtnnBfloat16:
-    def __repr__(self):
-        return f"ttnn_bfloat16"
 
 class AddDataMovePass(PassBase):
     def call(self, gm: torch.fx.GraphModule):
